refactor(bucket): Split code for lisibility
